TaskNotFoundException id has migrated to Long

+++ src/main/java/tk/hiddenname/smarthome/exception/TaskNotFoundException.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class TaskNotFoundException extends Exception {
 
-    public TaskNotFoundException(Integer id) {
+    public TaskNotFoundException(Long id) {
         super(String.format("The task with id '%d' not found", id));
     }
 }
